// src/api/generated/models/StickerResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */

export type StickerResponseDTO = {
    id: string;
    name: string;
    imageUrl: string;
    isAnimated: boolean;
    serverId: string;
    createdBy: string;
    createdAt?: string;
};